+++ test/test_kojo/subscribers/internal.subsA.js
@@ -9,4 +9,9 @@ module.exports = async (kojo, logger) => {
     assert(bravo.methodA);
     assert(typeof logger.debug === 'function');
     await sleep(15);
+    alpha.on('aCalled', (param) => {
+        assert(param === 'boo' );
+        logger.info('got alpha.aCalled with', param);
+        logger.debug('ending subscriber');
+    });
 };