+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+[4.2.0] - 2018-06-20
+--------------------
+- [x] Introduce `options.loglevel`
+- [x] Change logger to chalked loglevel
+- [x] Change `console.log` to `process.stdout.write` for splash screen
+- [x] Update dependencies
+
+
 [4.1.0] - 2018-05-31
 --------------------
 - [x] Add JSDoc